+++ src/main/kotlin/dev/d1s/linda/exception/AvailabilityCheckInProgressException.kt
@@ -14,10 +14,12 @@
  * limitations under the License.
  */
 
-package dev.d1s.linda.exception.impl.notFound
+package dev.d1s.linda.exception
 
-import dev.d1s.linda.exception.DomainNotFoundException
+import dev.d1s.advice.domain.ErrorResponseData
+import dev.d1s.advice.exception.HttpStatusException
+import org.springframework.http.HttpStatus
 
-object UtmParameterNotFoundException : DomainNotFoundException(
-    "UTM parameter was not found."
+object AvailabilityCheckInProgressException : HttpStatusException(
+    ErrorResponseData(HttpStatus.UNPROCESSABLE_ENTITY, "Availability check is already in progress.")
 )
