+++ src/main/kotlin/dev/d1s/linda/exception/notFound/DomainNotFoundException.kt
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *      https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,10 +14,12 @@
  * limitations under the License.
  */
 
-package dev.d1s.linda.exception.impl.notFound
+package dev.d1s.linda.exception.notFound
 
-import dev.d1s.linda.exception.DomainNotFoundException
+import dev.d1s.advice.domain.ErrorResponseData
+import dev.d1s.advice.exception.HttpStatusException
+import org.springframework.http.HttpStatus
 
-object AvailabilityChangeNotFoundException : DomainNotFoundException(
-    "Requested availability change was not found."
+open class DomainNotFoundException(message: String) : HttpStatusException(
+    ErrorResponseData(HttpStatus.NOT_FOUND, message)
 )
