+++ src/main/kotlin/dev/d1s/linda/exception/notFound/impl/RedirectNotFoundException.kt
@@ -14,15 +14,10 @@
  * limitations under the License.
  */
 
-package dev.d1s.linda.exception.impl.customAlias
+package dev.d1s.linda.exception.notFound.impl
 
-import dev.d1s.advice.domain.ErrorResponseData
-import dev.d1s.advice.exception.HttpStatusException
-import org.springframework.http.HttpStatus
+import dev.d1s.linda.exception.notFound.DomainNotFoundException
 
-object CustomAliasNotDefinedException : HttpStatusException(
-    ErrorResponseData(
-        HttpStatus.BAD_REQUEST,
-        "Custom alias is not defined. Specify it using the 'customAlias' request parameter."
-    )
+object RedirectNotFoundException : DomainNotFoundException(
+    "The requested redirect was not found."
 )
