+++ src/main/kotlin/dev/d1s/linda/exception/notFound/impl/ShortLinkNotFoundException.kt
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      https://www.apache.org/licenses/LICENSE-2.0
+ *     http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,12 +14,10 @@
  * limitations under the License.
  */
 
-package dev.d1s.linda.exception
+package dev.d1s.linda.exception.notFound.impl
 
-import dev.d1s.advice.domain.ErrorResponseData
-import dev.d1s.advice.exception.HttpStatusException
-import org.springframework.http.HttpStatus
+import dev.d1s.linda.exception.notFound.DomainNotFoundException
 
-open class DomainNotFoundException(message: String) : HttpStatusException(
-    ErrorResponseData(HttpStatus.NOT_FOUND, message)
+object ShortLinkNotFoundException : DomainNotFoundException(
+    "The requested short link was not found."
 )
