+++ src/main/kotlin/dev/d1s/linda/service/impl/AvailabilityChangeServiceImpl.kt
@@ -23,8 +23,8 @@ import dev.d1s.linda.domain.availability.AvailabilityChange
 import dev.d1s.linda.domain.availability.UnavailabilityReason
 import dev.d1s.linda.dto.availability.AvailabilityChangeDto
 import dev.d1s.linda.event.data.AvailabilityChangeEventData
-import dev.d1s.linda.exception.impl.AvailabilityCheckInProgressException
-import dev.d1s.linda.exception.impl.notFound.AvailabilityChangeNotFoundException
+import dev.d1s.linda.exception.AvailabilityCheckInProgressException
+import dev.d1s.linda.exception.notFound.impl.AvailabilityChangeNotFoundException
 import dev.d1s.linda.repository.AvailabilityChangeRepository
 import dev.d1s.linda.service.AvailabilityChangeService
 import dev.d1s.linda.service.ShortLinkService
