+++ src/main/kotlin/dev/d1s/linda/service/impl/UtmParameterServiceImpl.kt
@@ -19,8 +19,8 @@ package dev.d1s.linda.service.impl
 import dev.d1s.linda.domain.Redirect
 import dev.d1s.linda.domain.utm.UtmParameter
 import dev.d1s.linda.domain.utm.UtmParameterType
-import dev.d1s.linda.exception.impl.alreadyExists.UtmParameterAlreadyExistsException
-import dev.d1s.linda.exception.impl.notFound.UtmParameterNotFoundException
+import dev.d1s.linda.exception.alreadyExists.impl.UtmParameterAlreadyExistsException
+import dev.d1s.linda.exception.notFound.impl.UtmParameterNotFoundException
 import dev.d1s.linda.repository.UtmParameterRepository
 import dev.d1s.linda.service.RedirectService
 import dev.d1s.linda.service.UtmParameterService
