+++ src/test/kotlin/dev/d1s/linda/exception/customAlias/EmptyCustomAliasExceptionTest.kt
@@ -16,7 +16,7 @@
 
 package dev.d1s.linda.exception.customAlias
 
-import dev.d1s.linda.exception.impl.customAlias.EmptyCustomAliasException
+import dev.d1s.linda.exception.customAlias.impl.EmptyCustomAliasException
 import org.junit.jupiter.api.Test
 import strikt.api.expectThat
 import strikt.assertions.isEqualTo
