+++ src/test/kotlin/dev/d1s/linda/exception/notFound/AvailabilityChangeNotFoundExceptionTest.kt
@@ -16,7 +16,7 @@
 
 package dev.d1s.linda.exception.notFound
 
-import dev.d1s.linda.exception.impl.notFound.AvailabilityChangeNotFoundException
+import dev.d1s.linda.exception.notFound.impl.AvailabilityChangeNotFoundException
 import org.junit.jupiter.api.Test
 import strikt.api.expectThat
 import strikt.assertions.isEqualTo
